+++ eslint.config.mjs
@@ -1,39 +1,23 @@
-import globals from "globals";
-import babelParser from "@babel/eslint-parser";
-import path from "node:path";
-import { fileURLToPath } from "node:url";
-import js from "@eslint/js";
-import { FlatCompat } from "@eslint/eslintrc";
+import globals from 'globals'
+import js from '@eslint/js'
+import prettier from 'eslint-config-prettier'
 
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-const compat = new FlatCompat({
-    baseDirectory: __dirname,
-    recommendedConfig: js.configs.recommended,
-    allConfig: js.configs.all
-});
-
-export default [...compat.extends("eslint:recommended"), {
+export default [
+  {
+    ignores: ['**/package-lock.json', 'node_modules/**', '.release/**'],
+  },
+  js.configs.recommended,
+  prettier,
+  {
     languageOptions: {
-        globals: {
-            ...globals.node,
-        },
-
-        parser: babelParser,
-        ecmaVersion: "latest",
-        sourceType: "module",
-
-        parserOptions: {
-            babelOptions: {
-                configFile: false,
-                plugins: ["@babel/plugin-syntax-import-attributes"],
-            },
-
-            requireConfigFile: false,
-        },
+      ecmaVersion: 'latest',
+      sourceType: 'module',
+      globals: {
+        ...globals.node
+      },
     },
-
     rules: {
-        "no-unused-vars": "warn"
+      'no-unused-vars': 'warn',
     },
-}];
+  },
+]