+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ### Unreleased
 
+### [3.0.0-alpha.10] - 2026-03-25
+
+- config: replace .yaml with .toml
+- zone_record can be empty, default 0
+- feat(zone records): create and delete
+
 ### [3.0.0-alpha.9] - 2026-03-15
 
 - feat(zone): use DataTable for list, added search/limit options
@@ -64,3 +70,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 [3.0.0-alpha.7]: https://github.com/NicTool/api/releases/tag/v3.0.0-alpha.7
 [3.0.0-alpha.8]: https://github.com/NicTool/api/releases/tag/v3.0.0-alpha.8
 [3.0.0-alpha.9]: https://github.com/NicTool/api/releases/tag/v3.0.0-alpha.9
+[3.0.0-alpha.10]: https://github.com/NicTool/api/releases/tag/v3.0.0-alpha.10